requirements.txt
langchain_openai
langchain-core
streamlit
python-dotenv
langchain-community 
langchain-ollama
langchain
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
arxiv
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
jupyterlab


